refactor: extract ProgressMessage to messages.rs; make eframe optional behind gui feature
Prerequisite for WASM Component Model build (GH #68)
