+++ src/tsp/branch_bound.rs
@@ -3,7 +3,7 @@ use std::rc::Rc;
 
 use super::distance_matrix::DistanceMatrix;
 use super::kdtree::KDPoint;
-use super::progress::ProgressMessage;
+use super::messages::ProgressMessage;
 use super::route::Route;
 use super::{Solution, SolverOptions};
