+++ src/tsp/cuckoo_search.rs
@@ -3,7 +3,7 @@ use rand::Rng;
 use super::distance_matrix::DistanceMatrix;
 use super::kdtree::KDPoint;
 use super::probability::levy_step;
-use super::progress::ProgressMessage;
+use super::messages::ProgressMessage;
 use super::route::Route;
 use super::{Solution, SolverOptions};
 const DEFAULT_N_NESTS: usize = 25;
