+++ src/tsp/flower_pollination.rs
@@ -3,7 +3,7 @@ use rand::Rng;
 use super::distance_matrix::DistanceMatrix;
 use super::kdtree::KDPoint;
 use super::probability::{bernoulli, levy_step, sample_without_replacement};
-use super::progress::ProgressMessage;
+use super::messages::ProgressMessage;
 use super::route::{apply_swaps, swap_sequence, Route};
 use super::{Solution, SolverOptions};
