+++ src/tsp/nearest_neighbor.rs
@@ -2,7 +2,7 @@ use std::collections::{HashMap, HashSet};
 
 use super::distance_matrix::DistanceMatrix;
 use super::kdtree::KDPoint;
-use super::progress::ProgressMessage;
+use super::messages::ProgressMessage;
 use super::route::Route;
 use super::{Solution, SolverOptions};
