+++ src/tsp/particle_swarm.rs
@@ -2,7 +2,7 @@ use rand::Rng;
 
 use super::distance_matrix::DistanceMatrix;
 use super::kdtree::KDPoint;
-use super::progress::ProgressMessage;
+use super::messages::ProgressMessage;
 use super::route::{apply_swaps, swap_sequence, Route};
 use super::{Solution, SolverOptions};
