+++ src/tsp/two_opt.rs
@@ -1,6 +1,6 @@
 use super::distance_matrix::DistanceMatrix;
 use super::kdtree::KDPoint;
-use super::progress::ProgressMessage;
+use super::messages::ProgressMessage;
 use super::route::Route;
 use super::{Solution, SolverOptions};
